New backend setup

+++ app/src/main/java/com/example/luminara/data/model/Directory.kt
@@ -1,10 +1,10 @@
 package com.example.luminara.data.model
 
 data class Directory(
-    val id: Int,
+    val id: Long,
     val name: String,
     val address: String,
-    val overallRating: Int,
+    val overallRating: Float,
     val openingHours: String,
     val description: String,
     val mainImageUrl: String
